Move organization to flat routes
