+++ packages/insomnia/src/models/project.ts
@@ -1,7 +1,7 @@
 import { database as db } from '../common/database';
 import { generateId } from '../common/misc';
-import type { StorageRules } from '../ui/routes/organization';
 import { type BaseModel } from './index';
+import type { StorageRules } from './organization';
 
 export const name = 'Project';
 export const type = 'Project';
