+++ packages/insomnia/src/ui/components/dropdowns/project-dropdown.tsx
@@ -4,8 +4,8 @@ import { Button, Menu, MenuItem, MenuTrigger, Popover, Tooltip, TooltipTrigger }
 import { useFetcher } from 'react-router';
 
 import type { GitRepository } from '../../../models/git-repository';
+import type { StorageRules } from '../../../models/organization';
 import { getProjectStorageTypeLabel, isGitProject, isRemoteProject, type Project } from '../../../models/project';
-import { type StorageRules } from '../../routes/organization';
 import { Icon } from '../icon';
 import { showAlert, showModal } from '../modals';
 import { AskModal } from '../modals/ask-modal';
