+++ packages/insomnia/src/ui/components/modals/project-modal.tsx
@@ -3,8 +3,8 @@ import { Button, Dialog, Heading, Modal, ModalOverlay } from 'react-aria-compone
 import { useNavigation } from 'react-router';
 
 import type { GitRepository } from '../../../models/git-repository';
+import type { StorageRules } from '../../../models/organization';
 import type { Project } from '../../../models/project';
-import type { StorageRules } from '../../routes/organization';
 import { Icon } from '../icon';
 import { ProjectSettingsForm } from '../project/project-settings-form';
